requirements.txt
discord.py==1.4.1
numpy==1.19.1
opencv-python==4.3.0.36
Pillow==7.2.0
requests==2.22.0
APScheduler==3.6.3
python-dotenv==0.14.0